Move gammapy.time.utils to gammapy.utils.time.
This is needed for a clean gammapy package hierarchy,
where gammapy.utils.time is imported in gammapy.data,
which is lower-level than gammapy.time.

gammapy.time will be about time-based analysis for events and lightcurves.
