+++ gammapy/time/__init__.py
@@ -1,9 +1,8 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """
-Time handling and analysis.
+Time-based analysis.
 """
 from .plot import *
 from .simulate import *
-from .utils import *
 from .lightcurve import *
 from .exptest import *
