+++ gammapy/utils/time.py
